# unifor-backend/src/main/java/br/com/unifor/dto/DisciplineRequestDTO.java
package br.com.unifor.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;
import lombok.Data;

@Data
public class DisciplineRequestDTO {
    @NotBlank
    @Size(max = 20)
    private String code;

    @NotBlank
    @Size(max = 100)
    private String name;

    @NotNull
    private Integer credits;

    @NotNull
    private Integer workload;

    private String description;
}